refactor (factory): more descriptive folder structure
