+++ src/common/factory/journal/diary.rs
