+++ src/common/factory/journal/link.rs
