+++ src/common/factory/journal/mod.rs
@@ -1,0 +1,3 @@
+pub mod diary;
+pub mod link;
+pub mod reading_note;
