+++ src/common/factory/journal/reading_note.rs
