+++ src/common/factory/mod.rs
@@ -1,21 +1,15 @@
-mod action;
-mod action_track;
-mod ambition;
-mod desired_state;
-mod diary;
-mod link;
-mod mindset;
-mod reading_note;
+mod journal;
+mod my_way;
 mod tag;
 mod user;
 
-pub use action::*;
-pub use action_track::*;
-pub use ambition::*;
-pub use desired_state::*;
-pub use diary::*;
-pub use link::*;
-pub use mindset::*;
-pub use reading_note::*;
+pub use journal::diary::*;
+pub use journal::link::*;
+pub use journal::reading_note::*;
+pub use my_way::action::*;
+pub use my_way::action_track::*;
+pub use my_way::ambition::*;
+pub use my_way::desired_state::*;
+pub use my_way::mindset::*;
 pub use tag::*;
 pub use user::*;
