+++ src/common/factory/my_way/ambition.rs
