+++ src/common/factory/my_way/mindset.rs
