+++ src/common/factory/my_way/mod.rs
@@ -1,0 +1,5 @@
+pub mod action;
+pub mod action_track;
+pub mod ambition;
+pub mod desired_state;
+pub mod mindset;
